task & bug: Move constants to separate folder & Fix room and teacher validators
