+++ client/src/helpers/utilities.js
@@ -1,0 +1,18 @@
+// Check whether a cancellation took place on the same day of the class
+const checkForSameDate = (eventStart) => {
+  if (eventStart) {
+    const todaysDate = new Date();
+    return eventStart.getDate() === todaysDate.getDate();
+  } else {
+    return console.log("Event does not exist");
+  }
+};
+
+const protectAction = (user, action) => {
+  if (user.user.role === "staff") {
+    return action;
+  }
+  return console.log("Not gonna do it. Wouldn't be prudent.");
+};
+
+export { checkForSameDate, protectAction };
